backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pdfminer.six==20231228
python-docx==1.1.2
python-dotenv==1.0.1
google-genai==1.8.0